requirements.txt
asn1crypto==1.3.0
bitarray==1.2.1
certifi==2019.11.28
cffi==1.13.2
Click==7.0
cryptography==2.8
cx-Oracle==7.3.0
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.18.1
pandas==1.0.0
psycopg2==2.8.4
pybloom-mirror==2.0.0
pycparser==2.19
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
Werkzeug==0.16.1
wincertstore==0.2
